Removed the dummy page for the modal and linked it to the attendance-marking-page
