+++ src/routes/MainRoutes.js
@@ -4,8 +4,6 @@ import ErrorPage from '../pages/error-page';
 import ExamAttendanceReport from '../pages/exam_attendance_report';
 import AttendanceMarkingPage from '../pages/attendance-marking-page';
 import ExaminerLoginPage from '../pages/examiner-login-page';
-import StudentDetailsDummyPage from '../pages/student-details-dummy-page';
-
 
 const MainRoutes = {
   path: '/',
@@ -30,10 +28,6 @@ const MainRoutes = {
     {
       path: "attendance-marking",
       element: <AttendanceMarkingPage />
-    },
-    {
-      path: "student-details-dummy-page",
-      element: <StudentDetailsDummyPage />
     }
   ]
 };
